requirements.txt
streamlit
pandas
plotly
beautifulsoup4
requests
google-generativeai
python-dotenv
lxml
numpy
apify-client
textblob